Muted sanitizer errors on W32 platform.
Actually MinGW headers need to be fixed.

+++ src/examples/upgrade_example.c
@@ -57,7 +57,7 @@ make_blocking (MHD_socket fd)
 #elif defined(MHD_WINSOCK_SOCKETS)
   unsigned long flags = 1;
 
-  ioctlsocket (fd, FIONBIO, &flags);
+  ioctlsocket (fd, (int) FIONBIO, &flags);
 #endif /* MHD_WINSOCK_SOCKETS */
 
 }
