+++ src/examples/websocket_chatserver_example.c
@@ -708,7 +708,7 @@ make_blocking (MHD_socket fd)
 #elif defined(MHD_WINSOCK_SOCKETS)
   unsigned long flags = 0;
 
-  ioctlsocket (fd, FIONBIO, &flags);
+  ioctlsocket (fd, (int) FIONBIO, &flags);
 #endif /* MHD_WINSOCK_SOCKETS */
 
 }
